Move unrelated functions out of fetcher
