+++ pkg/services/ngalert/schedule/metrics_test.go
@@ -3,9 +3,8 @@ package schedule
 import (
 	"testing"
 
+	models "github.com/grafana/grafana/pkg/services/ngalert/models"
 	"github.com/stretchr/testify/assert"
-
-	"github.com/grafana/grafana/pkg/services/ngalert/models"
 )
 
 func TestHashUIDs(t *testing.T) {
